requirement.txt
flask
supabase-py
python-dotenv